ch(services/commands/*): move tests into one folder
