+++ services/commands/tests/delete.test.ts
@@ -1,7 +1,7 @@
 import test from 'tape'
-import deleteKudo from './delete'
-import utils from '../../test/utils'
-import kudoModel from '../../models/kudos'
+import deleteKudo from '../delete'
+import utils from '../../../test/utils'
+import kudoModel from '../../../models/kudos'
 
 test('before', async function (t) {
     t.plan(1)
